feat: 批量删除用户和岗位关联

+++ src/modules/system/repository/sys_user_post.go
@@ -8,7 +8,7 @@ type ISysUserPost interface {
 	CountUserPostByPostId(postId string) int
 
 	// DeleteUserPost 批量删除用户和岗位关联
-	DeleteUserPost(userIds []string) int
+	DeleteUserPost(userIds []string) int64
 
 	// BatchUserPost 批量新增用户岗位信息
 	BatchUserPost(sysUserPosts []model.SysUserPost) int
